% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{generate_data}
\alias{generate_data}
\title{Generate 2 timepoint data for RCI analysis}
\usage{
generate_data(
  n_sims = 1,
  n_ppl = 5000,
  sd_tot = 1,
  m_bl_tot = 0,
  delta = -0.5,
  delta_sd = 1,
  meas_err = 0.5
)
}
\arguments{
\item{n_sims}{Number of simulations}

\item{n_ppl}{Number of people to simulate}

\item{sd_tot}{Standard deviation of the true scores at baseline}

\item{m_bl_tot}{Mean of the true scores at baseline}

\item{delta}{Mean change, applied to all individuals uniformly.}

\item{delta_sd}{SD of change. Each person gets a value from this SD between time 0 and time 1.}

\item{meas_err}{Measurement error. SEm. SD of the error component at each time point.}
}
\value{
A tibble with simulated data and RCI-related variables.
}
\description{
Generate 2 timepoint data for RCI analysis
}
\examples{
generate_data()
}
