% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{generate_xt_data}
\alias{generate_xt_data}
\title{Generate X timepoint data for RCI analysis}
\usage{
generate_xt_data(
  n_sims = 1,
  n_ppl = 500,
  sd_tot = 1,
  m_bl_tot = 0,
  delta = -0.5,
  delta_sd = 1,
  meas_err = 0.5,
  n_obs
)
}
\arguments{
\item{n_sims}{Number of simulations}

\item{n_ppl}{Number of people to simulate}

\item{sd_tot}{Standard deviation of the true scores at baseline}

\item{m_bl_tot}{Mean of the true scores at baseline}

\item{delta}{Mean change per unit time, applied to all individuals uniformly.}

\item{delta_sd}{SD of change per unit time.}

\item{meas_err}{Measurement error (SEm). SD of the error component at each time point.}

\item{n_obs}{Number of observations per person}
}
\value{
A tibble with simulated longitudinal data and RCI-related variables.
}
\description{
Generate X timepoint data for RCI analysis
}
\examples{
generate_xt_data(n_obs = 10)
}
