% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{hist_obs_diff_raw}
\alias{hist_obs_diff_raw}
\title{Simple raw histogram of observed difference scores}
\usage{
hist_obs_diff_raw(x)
}
\arguments{
\item{x}{data containing columns labeled \code{time} and \code{obs_diff}.}
}
\value{
ggplot object
}
\description{
Simple raw histogram of observed difference scores
}
\examples{
hist_obs_diff_raw(generate_data())
}
