% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_reliableTrend.R
\name{is.reliableTrend}
\alias{is.reliableTrend}
\title{Is object a reliableTrend?}
\usage{
is.reliableTrend(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
A single logical value.
}
\description{
Predicate returning \code{TRUE} if \code{x} is an object of class
\code{"reliableTrend"} (as produced by \code{\link{rti}}), otherwise \code{FALSE}.
}
\examples{
fit <- rti(c(12, 14), sd = 5, r = 0.8)
is.reliableTrend(fit)     # TRUE
is.reliableTrend(list())  # FALSE

}
