% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jt_example_data_1}
\alias{jt_example_data_1}
\title{Jacobson & Truax (1991) data for a single case example}
\format{
A data.frame with 2 rows and 2 variables:
\describe{
\item{obs}{Observed score.}
\item{time}{Indicator for time.}
}
}
\source{
Jacobson, N. S. and Truax, P. (1991). Clinical significance: A statistical
approach to defining meaningful change in psychotherapy research. Journal of Consulting
and Clinical Psychology, 59, p. 12 - 19.
}
\usage{
jt_example_data_1
}
\description{
A \code{data.frame} containing the sample data for one patient provided in Jacobson & Truax (1991).
}
\keyword{datasets}
