% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{jt_rci_calc}
\alias{jt_rci_calc}
\title{Jacobson–Truax RCI from difference & SDiff}
\usage{
jt_rci_calc(difference, sdiff)
}
\arguments{
\item{difference}{observed pre–post difference (x2 - x1)}

\item{sdiff}{standard error of the difference score}
}
\value{
numeric RCI = difference / sdiff
}
\description{
Jacobson–Truax RCI from difference & SDiff
}
