% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liu_data}
\alias{liu_data}
\title{Liu Data}
\format{
A data.frame with 12 rows and 2 variables:
\describe{
\item{obs}{Observed score.}
\item{time}{Indicator for time.}
}
}
\source{
Liu, F. F., & Adrian, M. C. (2019). Is treatment working? Detecting real
change in the treatment of child and adolescent depression. Journal of the American
Academy of Child & Adolescent Psychiatry, 58(12), 1157-1164.
}
\usage{
liu_data
}
\description{
Data from a single case using the PHQ-9 Modified for Adolescents.
Presented in Liu, F. F., & Adrian, M. C. (2019). Is treatment working? Detecting real
change in the treatment of child and adolescent depression. Journal of the American
Academy of Child & Adolescent Psychiatry, 58(12), 1157-1164.
}
\details{
A \code{data.frame} containing the sample data for one individual measured at 12 times.
}
\keyword{datasets}
