% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mac_height}
\alias{mac_height}
\title{Mac height data}
\format{
A data.frame with 2 rows and 2 variables:
\describe{
\item{obs}{Observed score.}
\item{time}{Indicator for time.}
}
}
\source{
McAleavey 2022
}
\usage{
mac_height
}
\description{
Example data for a single case
}
\details{
A \code{data.frame} containing the sample data for one individual measured at six times.
}
\keyword{datasets}
