% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliableTrend_class.R
\name{new_reliableTrend}
\alias{new_reliableTrend}
\title{Basing on S3, following \href{https://adv-r.hadley.nz/s3.html#s3-classes}{Advanced R}.
Constructor function for reliableTrend class objects}
\usage{
new_reliableTrend(
  RCI = double(),
  RTI = double(),
  pd.RCI = double(),
  pd.RTI = double(),
  category.RTI = "Less than reliable",
  category.RCI = "Less than reliable",
  sign.RTI = "Not calculated",
  sign.difference = "Not calculated",
  rmaObj = list(),
  values = vector(mode = "numeric"),
  values.prepost = vector(mode = "numeric"),
  error_var = double(),
  cutpoint = double(),
  observed = "obs_score",
  scale_RCI = double()
)
}
\arguments{
\item{RCI}{Numeric. Jacobson & Truax RCI, a transformed difference score.}

\item{RTI}{Numeric. Reliable Trend Index, the z-value of a linear trend
considering a fixed measurement error at each observation.}

\item{pd.RCI}{Numeric.}

\item{pd.RTI}{Numeric.}

\item{category.RTI}{Character. Default is "Less than reliable". One of "Less
than reliable", "Reliable Increase", or "Reliable Decrease".}

\item{category.RCI}{Character. Default is "Less than reliable". One of "Less
than reliable", "Reliable Increase", or "Reliable Decrease".}

\item{sign.RTI}{Character. Indicates whether the RTI estimates the trend to
be most likely increasing, decreasing or flat, regardless of confidence
level. Possible values are \code{"Not calculated"} (the default), \code{"Increase"},
\code{"Decrease"}, or \code{"Flat"}.}

\item{sign.difference}{Character. Indicates whether the pre-post comparison
is increasing, decreasing or flat, based on simple subtraction of observed
scores. Possible values are \code{"Not calculated"} (the default), \code{"Increase"},
\code{"Decrease"}, or \code{"Flat"}.}

\item{rmaObj}{An object of type \code{rma} and/or \code{rma.uni}, derived from the
\code{{metafor}} package.}

\item{values}{Numeric vector. The values used to compute the RTI.}

\item{values.prepost}{Numeric vector of length 2. The first and last values.}

\item{error_var}{Numeric. The error variance used to compute the RCI and RTI.
This is on the variance scale, so is the squared standard error of the
difference.}

\item{cutpoint}{Numeric. Z-scale cutpoint to use for reliability
categorization. Default is 1.96.}

\item{observed}{Character. Name of the observed variable. Defaults to
\code{"obs_score"}.}

\item{scale_RCI}{Numeric. The "RCI for a scale," meaning how many scale
points need to be observed in a difference score for that difference to be
considered "reliable" under J&T.}
}
\value{
An object of type \code{reliableTrend}.
This object is a glorified list with the above-defined parameters as values.
}
\description{
Internal function to create objects that are of class \code{reliableTrend}.
This is an S3 class from the package \code{ReliableTrendIndex}.
Following \href{https://adv-r.hadley.nz/s3.html#s3-classes}{Advanced R}.
All reliableTrend objects use this (internally).
}
\examples{
new_reliableTrend()
}
\seealso{
\code{\link{reliableTrend}}
}
