% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{plot_rci_grps}
\alias{plot_rci_grps}
\title{Line plot with RCI groups}
\usage{
plot_rci_grps(x)
}
\arguments{
\item{x}{data from simulation}
}
\value{
ggplot object
}
\description{
Line plot with RCI groups
}
\examples{
plot_rci_grps(generate_data())
}
