% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{plotting_func_delta}
\alias{plotting_func_delta}
\title{Plotting simulation results from effect size simulation}
\usage{
plotting_func_delta(comparison.data)
}
\arguments{
\item{comparison.data}{A data frame, likely generated from bind_rows(into_output_tbl()).}
}
\description{
Plotting simulation results from effect size simulation
}
