% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.reliableTrend.R
\name{print.reliableTrend}
\alias{print.reliableTrend}
\title{Print method for reliableTrend objects}
\usage{
\method{print}{reliableTrend}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{A \code{reliableTrend} object (from \code{rti()}).}

\item{digits}{Number of significant digits to show.}

\item{...}{Ignored.}
}
\description{
Print method for reliableTrend objects
}
