% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeSErel.R
\name{print.slopeSErel}
\alias{print.slopeSErel}
\title{Print method for slopeSErel objects}
\usage{
\method{print}{slopeSErel}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"slopeSErel"} returned by
\code{\link{slope_se_reliability}}.}

\item{digits}{Number of digits to display.}

\item{...}{Further arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Pretty-prints the components of a \code{slopeSErel} object.
}
