% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCIcalculators.R
\name{rci_calc_jt}
\alias{rci_calc_jt}
\title{Calculate the Reliable Change Index according to Jacobson & Truax (1991).}
\usage{
rci_calc_jt(
  rxx = NULL,
  sd1 = NULL,
  sdiff = NULL,
  sem = NULL,
  prob = 0.975,
  verbose = FALSE
)
}
\arguments{
\item{rxx}{Numeric.  Reliability coefficient to use in calculation}

\item{sd1}{Numeric. Standard deviation 1. Should be a group representative of the target group. Calculated if not provided.}

\item{sdiff}{Numeric.  Standard error of the difference score. If provided, sem, rxx, sd1 are all ignored. Calculated if not provided.}

\item{sem}{Numeric. Standard error of measurement. Calculated if not provided.}

\item{prob}{Numeric 0-1. Defaults to .975 to provide 95\% two-sided confidence.}

\item{verbose}{Logical. If TRUE, will display additional information about the calculation.}
}
\value{
Numeric. The value of the RCI for the scale, meaning how large a change is required to be observed to be considered reliable.
}
\description{
Using a number of possible input combinations, most commonly time 1,
time 2, reliability, and SD value
}
\examples{
RCI1 <- rci_calc_jt(42, 30, .8, 15)
}
