% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCIcalculators.R
\name{RCIfunc}
\alias{RCIfunc}
\title{Calculate J&T's RCI on standardized scale}
\usage{
RCIfunc(rxx, s1 = 1, cut = 1.96)
}
\arguments{
\item{rxx}{The reliability coefficient}

\item{s1}{The standard deviation, Defaults to 1 for this example}

\item{cut}{The cutpoint, defaults to 1.96 to represent alpha = .05 two-sided, following J&T.}
}
\value{
The instrument-based RCI
}
\description{
This is just a simpler version of scale_rci_calc without the annoying parts and a slightly different order of events.
}
\examples{
RCIfunc(.9)
}
