% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{relChangePct_table}
\alias{relChangePct_table}
\title{Make a ReliableChange table by percentages}
\usage{
relChangePct_table(x, divideByTwo = TRUE)
}
\arguments{
\item{x}{data}

\item{divideByTwo}{logical.}
}
\value{
A table
}
\description{
Make a ReliableChange table by percentages
}
\examples{
relChangePct_table(generate_data())
}
