% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{relChangeRaw_table}
\alias{relChangeRaw_table}
\title{Generate a simple table of ReliableChange}
\usage{
relChangeRaw_table(x, divideByTwo = TRUE)
}
\arguments{
\item{x}{data}

\item{divideByTwo}{logical. Should the sample be divided by two
(e.g., if there are two lines per person)? Defaults to TRUE.}
}
\value{
A table.
}
\description{
Generate a simple table of ReliableChange
}
\examples{
relChangeRaw_table(generate_data())
relChangeRaw_table(generate_data(), divideByTwo = FALSE)
}
