% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliableTrend_class.R
\name{rti_to_stripped_list}
\alias{rti_to_stripped_list}
\title{Strip a \code{reliableTrend} object of non-atomic entries}
\usage{
rti_to_stripped_list(x)
}
\arguments{
\item{x}{An object of class \code{reliableTrend}.}
}
\value{
A list.
}
\description{
Takes an object of type \code{reliableTrend} and returns a list of atomic values.
}
\examples{
rti_to_stripped_list(reliableTrend())
}
