% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCIcalculators.R
\name{scale_rci_calc}
\alias{scale_rci_calc}
\title{Calculate the Reliable Change Index for a scale}
\usage{
scale_rci_calc(
  rxx = NULL,
  sd1 = NULL,
  sdiff = NULL,
  sem = NULL,
  prob = 0.975,
  verbose = FALSE
)
}
\arguments{
\item{rxx}{Numeric.  Reliability coefficient to use in calculation}

\item{sd1}{Numeric. Standard deviation 1. Should be a group representative of the target group. Calculated if not provided.}

\item{sdiff}{Numeric.  Standard error of the difference score. If provided, sem, rxx, sd1 are all ignored. Calculated if not provided.}

\item{sem}{Numeric. Standard error of measurement. Calculated if not provided.}

\item{prob}{Numeric 0-1. Defaults to .975 to provide 95\% two-sided confidence.}

\item{verbose}{Logical. If TRUE, will display additional information about the calculation.}
}
\value{
Numeric. The value of the RCI for the scale, meaning how large a change is required to be observed to be considered reliable.
}
\description{
This takes basic psychometric data and returns a value applicable
to difference scores, in theory. This is explicitly NOT what Jacobson & Truax (1991)
did, but it is common. The resulting value is the scale's RCI.
}
\details{
It is important to note that the \code{sdiff} value overwrites any other parameter if it is provided.
Typical inputs include:
sdiff with anything else; sem with anything else; rxx and sd1.
}
\examples{
RCI1 <- scale_rci_calc(42, 30, .8, 15)
}
