% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCIcalculators.R
\name{sdiff_calc}
\alias{sdiff_calc}
\title{Compute the standard error of the difference score}
\usage{
sdiff_calc(rxx, s1)
}
\arguments{
\item{rxx}{Reliability coefficient.}

\item{s1}{Standard deviation.}
}
\value{
Numeric value.
}
\description{
Only takes reliability and SD value per J&T (1991).
}
\examples{
sdiff_calc(.8, 7.5)
}
