% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{sdiff_from_sd_r}
\alias{sdiff_from_sd_r}
\title{SDiff from sd & r}
\usage{
sdiff_from_sd_r(sd, r)
}
\arguments{
\item{sd}{single-occasion SD}

\item{r}{reliability in \link{0,1}}
}
\value{
SDiff = sd * sqrt(2 * (1 - r))
}
\description{
SDiff from sd & r
}
