% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{sdiff_from_sem}
\alias{sdiff_from_sem}
\title{SDiff from SEM}
\usage{
sdiff_from_sem(sem)
}
\arguments{
\item{sem}{standard error of measurement}
}
\value{
SDiff = sqrt(2) * SEM
}
\description{
SDiff from SEM
}
