% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{sem_from_sd_r}
\alias{sem_from_sd_r}
\title{Standard Error of Measurement from sd & r}
\usage{
sem_from_sd_r(sd, r)
}
\arguments{
\item{sd}{single-occasion SD}

\item{r}{reliability in \eqn{\eqn{\eqn{[0, 1]}}}}
}
\value{
SEM = sd * sqrt(1 - r)
}
\description{
Standard Error of Measurement from sd & r
}
