% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliableTrend_class.R
\name{validate_reliableTrend}
\alias{validate_reliableTrend}
\title{Validate reliableTrend objects}
\usage{
validate_reliableTrend(x)
}
\arguments{
\item{x}{reliableTrend object.}
}
\value{
The same reliable trend object or errors.
}
\description{
An internal function
}
\examples{
validate_reliableTrend(new_reliableTrend())
}
\seealso{
\code{\link{reliableTrend}}
}
