% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuncs.R
\name{forest_to_reg_plot}
\alias{forest_to_reg_plot}
\title{Turning metafor's forest plots into regression plots with {ggplot2}}
\usage{
forest_to_reg_plot(x, StError, compare_lm_model = NULL)
}
\arguments{
\item{x}{A model from the {metafor} package. Designed only to work with univariate models.}

\item{StError}{Numeric. Standard error of the difference score to use.
Within the package should be available as the output of a call to rti_calulator()$variance}

\item{compare_lm_model}{Either NULL (default) or an object from lm().}
}
\value{
a ggplot2 object
}
\description{
Turning metafor's forest plots into regression plots with {ggplot2}
}
\examples{
test <- rti_calc_simple(c(47.5, 32.5), 4.74^2)
forest_to_reg_plot(test$rmaObj, StError = sqrt(test$variance)) +
  labs(title = "The example data from J&T")

output2 <- rti_calc_simple(c(98,98,98,99,99,99), .7071068^2)
forest_to_reg_plot(output2$rmaObj, StError = sqrt(output2$error_var)) +
  labs(title = "Mac's height example", 
       y = "Height difference from baseline")
}
