% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuncs.R
\name{forest_to_reg_plot}
\alias{forest_to_reg_plot}
\title{Turning metafor's forest plots into regression plots with {ggplot2}}
\usage{
forest_to_reg_plot(x, StError = SEm, compare_lm_model = NULL)
}
\arguments{
\item{x}{A model from the {metafor} package. Designed only to work with univariate models.}

\item{StError}{Numeric. Standard error of the difference score to use.
Defaults to \code{SEm} which WILL cause problems.  Within the package should be available as the output of a call to rti_calulator()$variance}

\item{compare_lm_model}{Either NULL (default) or an object from lm().}
}
\value{
a ggplot2 object
}
\description{
Turning metafor's forest plots into regression plots with {ggplot2}
}
\examples{
forest_to_reg_plot(MacMeta)
}
