% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuncs.R
\name{forest_to_reg_plot}
\alias{forest_to_reg_plot}
\title{Turning metafor's forest plots into regression plots with {ggplot2}}
\usage{
forest_to_reg_plot(x, StError = NULL, compare_lm_model = NULL)
}
\arguments{
\item{x}{A model from the {metafor} package or an object of class = \code{reliableTrend}.
Designed only to work with univariate models.}

\item{StError}{Numeric. Standard error of the measurement (\code{sem}, not \code{sdiff}).
If \code{x} is class \code{reliableTrend}, it is computed automatically.}

\item{compare_lm_model}{Either NULL (default) or an object from lm().}
}
\value{
a ggplot2 object
}
\description{
Turning metafor's forest plots into regression plots with {ggplot2}
}
\examples{
test <- rti_calc_simple(c(47.5, 32.5), 3.35)
forest_to_reg_plot(test$rmaObj, StError = sqrt(test$variance)) +
  labs(title = "The example data from J&T")

output2 <- rti_calc_simple(c(98,98,98,99,99,99), .5)
forest_to_reg_plot(output2$rmaObj, StError = output2$error_var) +
  labs(title = "Mac's height example", 
       y = "Height difference from baseline")
# test with reliableTrend object
forest_to_reg_plot(rti(mac_height$obs, sem = .5))
}
