% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{hist_obs_diff_raw}
\alias{hist_obs_diff_raw}
\title{Simple raw histogram of observed difference scores}
\usage{
hist_obs_diff_raw(x)
}
\arguments{
\item{x}{}
}
\description{
Simple raw histogram of observed difference scores
}
