% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliableTrend_class.R
\name{is.reliableTrend}
\alias{is.reliableTrend}
\title{Verify object claims to be \code{reliableTrend} class}
\usage{
is.reliableTrend(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
Logical.
}
\description{
Verify object claims to be \code{reliableTrend} class
}
\examples{
is.reliableTrend(reliableTrend())
is.reliableTrend("NOPE")
}
