% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCIcalculators.R
\name{jt_rci_calc}
\alias{jt_rci_calc}
\title{Compute the RCI value as in Jacobson & Truax (1991).}
\usage{
jt_rci_calc(
  difference = NULL,
  t1 = NULL,
  t2 = NULL,
  scale_rci = NULL,
  rxx = NULL,
  sd1 = NULL,
  sdiff = NULL,
  sem = NULL,
  prob = 0.975,
  verbose = FALSE
)
}
\arguments{
\item{difference}{}

\item{t1}{}

\item{t2}{}

\item{scale_rci}{}

\item{rxx}{}

\item{sd1}{}

\item{sdiff}{}

\item{sem}{}

\item{prob}{}

\item{verbose}{}
}
\description{
This provides one RCI value for every observed difference score,
essentially intended to be interpreted as Z-scores.
}
\examples{
jt_rci_calc(difference = 15, sdiff = 4.74)
}
