% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuncs.R
\name{plot.reliableTrend}
\alias{plot.reliableTrend}
\title{Plot a reliableTrend object}
\usage{
\method{plot}{reliableTrend}(x)
}
\arguments{
\item{x}{An object of class \code{reliableTrend}.}
}
\value{
A ggplot2 object
}
\description{
Plot a reliableTrend object
}
\examples{
mac_rti <- rti(mac_height$obs, sdiff = .707)
plot(mac_rti)
}
