% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliableTrend_class.R
\name{print.reliableTrend}
\alias{print.reliableTrend}
\title{Print a \code{reliableTrend} object}
\usage{
\method{print}{reliableTrend}(x, ...)
}
\arguments{
\item{x}{An object of class \code{reliableTrend}.}

\item{...}{Additional arguments.}
}
\value{
(invisibly) \code{x}.
}
\description{
Print a \code{reliableTrend} object
}
\examples{
output <- rti_calc_simple(c(98,98,98,99,99,99), .7071068^2)
print(output)
}
