% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rci.R
\name{rci}
\alias{rci}
\title{Compute the RCI value as in Jacobson & Truax (1991) and/or alternative
two-observation estimands.}
\usage{
rci(
  difference = NULL,
  t1 = NULL,
  t2 = NULL,
  scale_rci = NULL,
  r1 = NULL,
  r2 = NULL,
  sd1 = NULL,
  sd2 = NULL,
  sdiff = NULL,
  sem = NULL,
  prob = 0.975,
  verbose = FALSE,
  rc.type = "jt"
)
}
\arguments{
\item{difference}{Numeric. Difference score(s). Computed from \code{t1} and \code{t2} if not provided.}

\item{t1}{Numeric. Time 1 value(s). Ignored if \code{difference} is provided.}

\item{t2}{Numeric. Time 2 value(s). Ignored if \code{difference} is provided.}

\item{scale_rci}{Numeric. RCI for the scale, meaning the smallest change considered reliable.}

\item{r1}{Numeric.  First (or only) reliability coefficient to use in calculation}

\item{r2}{Numeric.  Optional second reliability coefficient to use in calculation}

\item{sd1}{Numeric. Optional standard deviation 2. Should be a group representative of the target group. Calculated if not provided.}

\item{sdiff}{Numeric.  Standard error of the difference score. If provided, sem, rxx, sd1 are all ignored. Calculated if not provided.}

\item{sem}{Numeric. Standard error of measurement. Calculated if not provided.}

\item{prob}{Numeric 0-1. Defaults to .975 to provide 95\% two-sided confidence.}

\item{verbose}{Logical. If TRUE, will return additional information about the calculation.}
}
\value{
Individual RCI value per difference score provided.
}
\description{
This provides one RCI value for every observed difference score,
essentially intended to be interpreted as Z-scores.
}
\details{
While the Jacobson & Truax (1991) definition (which is actually the Christensen & Mendoza,
1986 correction) has become the dominant version of providing statistical tests to the
difference between two scores from one person, it is merely the simplest possible
version of this estimand. To achieve simplicity, it relies on assumptions that will not
generally be true, but you may be able to relax by including more information.

For instance, it is possible that the SD among a group of individuals at time 1 is different
than the SD at time 2, for instance because treatment will de-homogenize individuals or
induce greater heterogeneity depending on the circumstance. Additionally, the reliability
coefficient may be meaningfully different at the two timepoints. Accommodating these
possibilities is not difficult when computing the RCI, and will only serve to make
the assumptions less wildly unjustified.
}
\examples{
jt_rci_calc(difference = 15, sdiff = 4.74)
jt_rci_calc(difference = 1, sdiff = .707)
}
