% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCIcalculators.R
\name{rci_classifier}
\alias{rci_classifier}
\title{Classify scores using Reliable Change Index values}
\usage{
rci_classifier(difference = NULL, scale_rci = NULL)
}
\arguments{
\item{difference}{Difference score(s)}

\item{scale_rci}{Scale RCI value}
}
\value{
A character vector with the standard values of "Reliably Increased", "Reliably Decreased", or "No Reliable Difference".
Also prints a warning to screen indicating that this is probably a terrible idea
}
\description{
This function assumes you only want categories of reliable change based on
Jacobson & Truax (1991). You probably don't want this.
}
\examples{
rci_classifier(difference = c(5, 3, -10), scale_rci = 4)
}
