% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliableTrend_class.R
\name{reliableTrend}
\alias{reliableTrend}
\title{Create a new reliableTrend object}
\usage{
reliableTrend(
  x = NULL,
  RCI = double(),
  RTI = double(),
  pd.RCI = double(),
  pd.RTI = double(),
  category.RTI = "Unspecified",
  category.RCI = "Unspecified",
  sign.RTI = "Not calculated",
  sign.difference = "Not calculated",
  rmaObj = list(),
  values = vector(mode = "numeric"),
  values.prepost = vector(mode = "numeric"),
  error_var = double(),
  cutpoint = 1.96,
  scale_RCI = double(),
  observed = "obs_score"
)
}
\arguments{
\item{x}{An object, likely with class \code{list} or \code{rma} (or \code{rma.uni}). If \code{x} is
provided, several other arguments are overwritten even if they are provided.}

\item{RCI}{Numeric. Jacobson & Truax RCI, a transformed difference score.}

\item{RTI}{Numeric. Reliable Trend Index, the z-value of a linear trend considering a
fixed measurement error at each observation.}

\item{pd.RCI}{Numeric.}

\item{pd.RTI}{Numeric.}

\item{category.RTI}{Character. Default is "Unspecified". One of "Unspecified",
"Reliable Increase", or "Reliable Decrease".}

\item{category.RCI}{Character. Default is "Unspecified". One of "Unspecified",
"Reliable Increase", or "Reliable Decrease".}

\item{sign.RTI}{Character. Indicates whether the RTI estimates the trend to be
most likely increasing, decreasing or flat, regardless of confidence level.
Possible values are \code{"Not calculated"} (the default), \code{"Increase"}, \code{"Decrease"}, or \code{"Flat"}.}

\item{sign.difference}{Character. Indicates whether the pre-post comparison
is increasing, decreasing or flat, based on simple subtraction of observed scores.
Possible values are \code{"Not calculated"} (the default), \code{"Increase"}, \code{"Decrease"}, or \code{"Flat"}.}

\item{rmaObj}{An object of type \code{rma} and/or \code{rma.uni}, derived from the \code{{metafor}}
package.}

\item{values}{Numeric vector. The values used to compute the RTI.}

\item{values.prepost}{Numeric vector of length 2. The first and last values.}

\item{error_var}{Numeric. The error variance used to compute the RCI and RTI.
This is on the variance scale, so is the squared standard error of the difference.}

\item{cutpoint}{Numeric. Z-scale cutpoint to use for reliability categorization.
Default is 1.96.}

\item{scale_RCI}{Numeric. The "RCI for a scale," meaning how many scale points
need to be observed in a difference score for that difference to be considered
"reliable" under J&T.}

\item{observed}{Character. Name of the variable of data frame \code{x} with the observed scores.}
}
\value{
an object of class \code{reliableTrend}, which is a glorified list.

To print a very simple summary use print(). For the full values, use summary().
}
\description{
This exported function creates \code{reliableTrend} objects using either direct input,
list input, or a standard \code{rma} object from \code{{metafor}}.
}
\examples{
output2 <- rti_calc_simple(c(98,98,98,99,99,99), .7071068^2)
reliableTrend(output2$rmaObj)
reliableTrend()
}
