% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReliableTrendIndex-package.R
\docType{package}
\name{ReliableTrendIndex-package}
\alias{ReliableTrendIndex}
\alias{ReliableTrendIndex-package}
\title{ReliableTrendIndex: Utilities for (More Reliable) Reliable Change Analysis}
\description{
You should not use this package because it makes RCI easy. The package enables simple calculations in a safe and consistent format, with the intent to encourage users to opt for increasingly reliable change analysis.
}
\author{
\strong{Maintainer}: Andrew McAleavey \email{andrew.mcaleavey@gmail.com} (\href{https://orcid.org/0000-0001-5986-2033}{ORCID})

}
\keyword{internal}
