% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTI.R
\name{rti}
\alias{rti}
\title{Compute RTI in a simple way}
\usage{
rti(
  values,
  sdiff = NULL,
  sem = NULL,
  scale_rci = NULL,
  cutpoint = 1.96,
  sd = NULL,
  rxx = NULL
)
}
\arguments{
\item{values}{}

\item{sdiff}{}

\item{sem}{}

\item{scale_rci}{}

\item{sd}{}

\item{rxx}{}
}
\value{
An object of class \code{reliableTrend}.
}
\description{
Compute RTI in a simple way
}
\examples{
rti(jt_example_data_1$obs, sdiff = 4.74)
rti(mac_height$obs, sdiff = .707)
rti(jt_example_data_1$obs, sem = 3.35)
rti(mac_height$obs, sem = .5)
rti(jt_example_data_1$obs, scale_rci = 9.2904)
rti(mac_height$obs, scale_rci = 1.385)
rti(jt_example_data_1$obs, sd = 7.5, rxx = .8)
rti(mac_height$obs, sd = 1.02, rxx = .77)
need one of these:
1 sdiff
2 sem
3 scale_rci
4 sd and rxx
}
