% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTI.R
\name{rti_calc_simple}
\alias{rti_calc_simple}
\title{Compute the RCI and RTI for a single person's observations.}
\usage{
rti_calc_simple(
  values,
  variance,
  digits = 2,
  cutpoint = 1.96,
  fixIntWhen3 = TRUE,
  ...
)
}
\arguments{
\item{values}{Numeric. Either a single difference value, two observations, or
more than two observations of the same variable for one person. While the
three cases are treated separately under the hood and return slightly
different values and text, the computation is the same and all involve
primarily \code{metafor::rma.uni()}.}

\item{variance}{Numeric. In most cases, this should be the squared standard
error of measurement (SEm). This is the error variance of a given
observation. NOT the Sdiff directly, but if you have the Sdiff previously
computed, you can divide it by the square root of 2. However, if only
providing a single observation, this is assumed to be a difference score,
so in that case, \code{variance} should be the square of Sdiff.}

\item{digits}{Integer. Number of digits to print for metafor::rma(). Defaults
to 2.}

\item{cutpoint}{Cutpoint on z-scale to use for "reliability." Defaults to
1.96.}

\item{fixIntWhen3}{Logical. Should the intercept be fixed when three
observations are provided? Defaults to true, which is similar to the
two-timepoint RCI solution (which assumes both observations are fixed).
Setting to false will mean the RTI and RCI are identical to each other with
3 observations, since there are no degrees of freedom in this model, so the
linear trend cannot be estimated.}

\item{...}{Additional arguments passed on.}
}
\value{
A list including the JT RCI value per observation in difs, A
classification per observation in difs
}
\description{
This is essentially a wrapper for \code{metafor::rma()} with some specific
behavior. It takes one or more observations with an indication of error
variance, and computes the RCI and/or RTI as applicable.
}
\details{
If only one value is provided, it assumes this is a difference score, and the
\code{variance} should be the square of Sdiff. If two or more values are provided,
it is assumed that these are raw observations, so \code{variance} should be the
square of SEm.

In all cases with three or more observations, this will assume observations
are evenly spaced in time (with a message).
}
\examples{
output <- rti_calc_simple(c(47.5, 32.5), 3.35^2)
output
metafor::forest(output$rmaObj)
output2 <- rti_calc_simple(c(5, 4, 3, 2), 1)
output2
metafor::forest(output2$rmaObj)
summary(output2$rmaObj)
forest_to_reg_plot(output2$rmaObj, StError = output2$error_var)

}
