% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTI.R
\name{rti_calc_simple}
\alias{rti_calc_simple}
\title{Compute the RCI and RTI for a single person's observations..}
\usage{
rti_calc_simple(values, variance, digits = 2, cutpoint = 1.96, ...)
}
\arguments{
\item{values}{Numeric. Either a single difference value, two observations, or more
than two observations of the same variable for one person. While the three cases are
treated separately under the hood and return slightly different values and text,
the computation is the same and all involve primarily \code{metafor::rma.uni()}.}

\item{variance}{Numeric. The  SEm!!!!! The error variance of a given observation.
NOT the Sdiff directly, but if you have
the Sdiff previously computed, you can divide it by the square root of 2.}

\item{digits}{Integer. Number of digits to print for metafor::rma(). Defaults to 2.}

\item{cutpoint}{Cutpoint on z-scale to use for "reliability." Defaults to 1.96.}

\item{...}{Additional arguments passed on.}
}
\value{
A list including the JT RCI value per observation in difs,
A classification per observation in difs
}
\description{
This is essentially a wrapper for \code{metafor::rma()} with some specific behavior.
It takes one or more observations with an indication of error variance, and computes
the RCI and/or RTI as applicable.
}
\examples{
output <- rti_calc_simple(c(47.5, 32.5), 4.74^2)
output
metafor::forest(output$rmaObj)
output2 <- rti_calc_simple(c(5, 4, 3, 2), 1)
output2
metafor::forest(output2$rmaObj)
summary(output2$rmaObj)
forest_to_reg_plot(output2$rmaObj, StError = output2$variance)

}
