% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{rxx_empirical}
\alias{rxx_empirical}
\title{Compute reliability from known true and observed scores}
\usage{
rxx_empirical(x)
}
\arguments{
\item{x}{data}
}
\value{
empirical reliability from a data set for simulation
}
\description{
Compute reliability from known true and observed scores
}
\examples{
rxx_empirical(generate_data())
}
