% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliableTrend_class.R
\name{summary.reliableTrend}
\alias{summary.reliableTrend}
\title{Text summary of a \code{reliableTrend} object}
\usage{
\method{summary}{reliableTrend}(x)
}
\arguments{
\item{x}{A \code{reliableTrend} object}
}
\value{
(invisibly) \code{x}.
}
\description{
Text summary of a \code{reliableTrend} object
}
\examples{
output <- rti_calc_simple(c(98,98,98,99,99,99), .7071068^2)
# gives a warning:
summary(output)
# compute the missing values using this:
output2 <- reliableTrend(output$rmaObj)
# no warning:
summary(output2)
}
